bender_service/requirements.txt
Django==1.10.4
Markdown==2.6.7
certifi==2017.7.27.1
coverage==4.3.4
django-access-logger==0.4.2
django-allauth==0.30.0
django-cors-headers==1.3.1
django-crispy-forms==1.6.1
django-extensions==1.7.5
django-filter==1.0.1
django-nose==1.4.4
django-rest-auth==0.9.0
djangorestframework-jwt==1.9.0
djangorestframework==3.5.3
gevent==1.2.1
gunicorn==19.6.0
ipdb==0.10.2
ipython==5.1.0
mock==2.0.0
numpy==1.12.0
psycopg2==2.6.2
python-json-logger==0.1.8
raven==5.31.0
scipy==0.18.1
setproctitle==1.1.10